+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changelog
 * Tools: fixed re-importing existing units sometimes losing component data.
 * Tools: fixed imported components not being removed from units when disabled on re-import.
 * Tools: Android: packages now set min SDK 24 and target SDK 34 in generated manifests.
+* Tools: Android: fixed deploying with debug keystore.
 * Docs: updated Android deploying instructions.
 * Docs: updated Importing Scenes with additional info on preparing models for export from DCCs.